Se agrega la documentación restante de los ejercicios del proyecto 2

+++ firmware/projects/guia2_ej3/main/guia2_ej3.c
@@ -1,8 +1,11 @@
-/*! @mainpage guia2_ej2
+/*! @mainpage guia2_ej3
  *
  * @section genDesc General Description
  *
- * This section describes how the program works.
+ * Este programa recibe de un dispositivo emisor de ultrasonido HC_SR04 el valor de distancia hasta la pared mas cercana donde
+ * rebota la onda emitida y posteriormente es recibida por el dispositio transformando el delay de llegada en un valor de
+ * distancia en cm.  En este caso el programa realizará sus acciones por medio de tareas e interrupciones y es capaz de enviar
+ * a travez de formato UART el valor de distancia obtenido a partir del dispositivo HC_SR04.
  *
  * <a href="https://drive.google.com/...">Operation Example</a>
  *
@@ -12,6 +15,13 @@
  * |:--------------:|:--------------|
  * | 	GPIO_3	 	| 	ECHO		|
  * | 	GPIO_2	 	| 	TRIGGER		|
+ * | 	D1		 	| 	GPIO_20		|
+ * | 	D2		 	| 	GPIO_21		|
+ * | 	D3		 	| 	GPIO_22		|
+ * | 	D4		 	| 	GPIO_23		|
+ * | 	SEL_1	 	| 	GPIO_19		|
+ * | 	SEL_2	 	| 	GPIO_18		|
+ * | 	SEL_3	 	| 	GPIO_9		|
  * | 	+5V	 	| 	+5V		|
  * | 	GND	 	| 	GND		|
  *
@@ -41,18 +51,34 @@
 /*==================[macros and definitions]=================================*/
 #define CONFIG_REFRESH_PERIOD_DISTANCE_US 1000000
 /*==================[internal data definition]===============================*/
-uint32_t distancia = 0;
+/* Entero de 16 bits donde se almacenará la distancia obtenida por el HC_SR04
+ */
+uint16_t distancia = 0;
+/* Booliano que almacena true si se actualiza la medicion o false si no se actualiza.
+ */
 bool medir = true;
+/* Booliano que almacena false si se actualiza la medicion en el display o true si no se actualiza
+ * la medicion en el display.
+ */
 bool hold = false;
 TaskHandle_t medir_task_handle = NULL;
 TaskHandle_t mostrar_task_handle = NULL;
 TaskHandle_t teclas_task_handle = NULL;
 /*==================[internal functions declaration]=========================*/
+/** @fn static void funcionTimerA(void *pvParameter)
+ * @brief Funcion que contiene el código que se ejecuta cada vez que se actualiza el timer A,
+ * el cual envia una notificación para realizar tareas.
+ * @return
+ */
 static void funcionTimerA(void *pvParameter)
 {
 	vTaskNotifyGiveFromISR(medir_task_handle, pdFALSE);
 }
-
+/** @fn static void medirDistancia(void *pvParameter)
+ * @brief Funcion que actualiza el valor de medición y envia una notificación a la funcion
+ * mostrarDistancia().
+ * @return
+ */
 static void medirDistancia(void *pvParameter)
 {
 	while (1)
@@ -65,7 +91,11 @@ static void medirDistancia(void *pvParameter)
 		vTaskNotifyGiveFromISR(mostrar_task_handle, pdFALSE);
 	}
 }
-
+/** @fn static void mostrarDistancia(void *pvParameter)
+ * @brief Funcion que actualiza el valor de medición mostrado en el display lcd, enciende los leds y
+ * envia en formato UART el valor de distancia almacenada.
+ * @return
+ */
 static void mostrarDistancia(void *pvParameter)
 {
 	while (1)
@@ -100,17 +130,27 @@ static void mostrarDistancia(void *pvParameter)
 		UartSendString(UART_PC, " cm\r\n");
 	}
 }
-
+/** @fn static void funcionTecla1(void *pvParameter)
+ * @brief Funcion que contiene la acción a ejecutar en caso de presionar el boton 1
+ * @return
+ */
 static void funcionTecla1(void *pvParameter)
 {
 	medir = !medir;
 }
-
+/** @fn static void funcionTecla2(void *pvParameter)
+ * @brief Funcion que contiene la acción a ejecutar en caso de presionar el boton 2
+ * @return
+ */
 static void funcionTecla2(void *pvParameter)
 {
 	hold = !hold;
 }
-
+/** @fn void funcionUARTRead(void *pvParameter)
+ * @brief Función que lee del dispositivo que se este enviado y recibiendo valores en formato UART
+ * un caracter para modificar el funcionamiento del programa.
+ * @return
+ */
 void funcionUARTRead(void *pvParameter)
 {
 	uint8_t caracter;
@@ -128,6 +168,11 @@ void funcionUARTRead(void *pvParameter)
 	}
 }
 
+/** @fn void inits()
+ * @brief Funcion que inicializa todos los dispositivos necesarios para el funcionamiento del programa
+ * exceptiando el timer y UART.
+ * @return
+ */
 void inits()
 {
 	HcSr04Init(GPIO_3, GPIO_2);
